#121 moved files to /messages/response and /message/answer
