+++ src/digitalIdentity-backend/src/main/java/didentity/amos/digitalIdentity/messages/answers/Accesstoken.java
@@ -1,4 +1,4 @@
-package didentity.amos.digitalIdentity.messages;
+package didentity.amos.digitalIdentity.messages.answers;
 
 import java.util.HashMap;
 import java.util.Map;
